ai-engine/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
redis[asyncio]>=5.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
numpy>=1.26.0
yfinance>=0.2.38
ta>=0.11.0
google-genai>=1.11.0
google-generativeai>=0.8.0
langgraph>=0.2.0
transformers>=4.44.0
torch>=2.3.0
typing-extensions>=4.12.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
